add manager object
This implements a mainloop that monitors firmware devices and listens
for signals.

The manager subscribes to and enumerates all firmware devices, and
processes any firmware requests from the kernel. It keeps a list of
firmware directories, and it tries to find the firmware file with
the correct name for any request that comes in. If found, this is
copied to the kernel, and if nothing is found, the request is
cancelled.