add firmware loading library
The firmware loading library operate on a dirfd representing the
device directory in /sys of the firmware device. It implements only
two funcions, firmware_load() and firmware_cancel_load().

The former takes an fd of the firmware file and copies this into the
kernel. The latter indicate to the kernel that the firmware cannot
be loaded.