allow firmwared to be run in 'tentative' mode
This means that a firmware request is only cancelled if something goes
wrong with the actual writing of the firmware to the kernel. Any other
errors are ignored, and the firmware request is left pending.

The idea is that this can be used when running firmwared before all
firmware is known to be available, under the assumption that firmwared
will be ran again once all firmware is ready.

Tentative mode would typicall be used in an initrd.